Extract MediaComponent tabs column to MediaComponentRightPanel component
In order to simplify MediaComponent, and put an ErrorBoundary specific for the right-side panel.

Fixes CHECK-1788
