+++ localization/react-intl/src/app/components/media/MediaOtherSidebar.json
@@ -1,0 +1,18 @@
+[
+  {
+    "id": "mediaComponent.requests",
+    "defaultMessage": "Requests"
+  },
+  {
+    "id": "mediaComponent.annotation",
+    "defaultMessage": "Annotation"
+  },
+  {
+    "id": "mediaComponent.source",
+    "defaultMessage": "Source"
+  },
+  {
+    "id": "mediaComponent.notes",
+    "defaultMessage": "Notes"
+  }
+]
